created useAdminUser hooks
